Fixing a bug for function definition
The function defintion should start one posiiton after the end of the modifier.
Given a function "function method() onlyOwner {...}', the current version of the return the definition "r{...}" instead of "{...}". I notice the same error happens when a function uses a modifier: the last character of the modifier is added to the  parsed definiton.

This commit fixes the bug.